requirements.txt
pyside6==6.9.1
pyqtgraph==0.13.7
pyserial==3.4